feat: Refactor imports in client and command files for better structure
